Test `FeedItemsMapper` in isolation
