+++ EssentialFeed/Feed API/RemoteFeedLoader.swift
@@ -11,6 +11,6 @@ public typealias RemoteFeedLoader = RemoteLoader<[FeedImage]>
 
 public extension RemoteFeedLoader {
     convenience init(url: URL, client: HTTPClient) {
-        self.init(url: url, client: client, mapper: FeedItemMapper.map)
+        self.init(url: url, client: client, mapper: FeedItemsMapper.map)
     }
 }
